implement another deprecation pattern
