+++ zingolib/Cargo.toml
@@ -7,7 +7,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-integration-tests = []
+deprecations = ["lightclient-deprecated"]
+lightclient-deprecated = []
 default = ["embed_params"]
 embed_params = []
